Move email handling to event handler
